+++ src/components/Activities/ActivityItem.css
@@ -1,4 +1,4 @@
-.fitness-item {
+.activity-item {
   display: flex;
   justify-content: space-between;
   align-items: center;
@@ -7,7 +7,7 @@
   background-color: #4b4b4b;
 }
 
-.fitness-item__description {
+.activity-item__description {
   display: flex;
   flex-direction: column;
   gap: 1rem;
@@ -17,14 +17,14 @@
   flex: 1;
 }
 
-.fitness-item h2 {
+.activity-item h2 {
   font-size: 1rem;
   flex: 1;
   margin: 0 1rem;
   color: white;
 }
 
-.fitness-item__minutes {
+.activity-item__minutes {
   font-size: 1rem;
   font-weight: bold;
   color: white;
@@ -35,18 +35,18 @@
 }
 
 @media (min-width: 580px) {
-  .fitness-item__description {
+  .activity-item__description {
     flex-direction: row;
     align-items: center;
     justify-content: flex-start;
     flex: 1;
   }
 
-  .fitness-item__description h2 {
+  .activity-item__description h2 {
     font-size: 1.25rem;
   }
 
-  .fitness-item__minutes {
+  .activity-item__minutes {
     font-size: 1.25rem;
     padding: 0.5rem 1.5rem;
   }
